refactor: extract workspacecreatedialog to own component
